# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Contoh API documentation",
        "description": "Contoh API doc menggunakan OpenAPI/Swagger",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "choirudin.emchagmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1"
    },
    "paths": {
        "/api/getgallery": {
            "get": {
                "tags": [
                    "Get Data Gallery"
                ],
                "summary": "Untuk Mendapatkan Data Gallery",
                "description": "Tes apakah API Gallery berjalan?",
                "operationId": "GetGallery",
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/postGallery": {
            "post": {
                "tags": [
                    "Upload Gambar"
                ],
                "summary": "Mengunggah Gambar",
                "description": "Endpoint yang digunakan untuk mengunggah gambar.",
                "operationId": "postGallery",
                "requestBody": {
                    "description": "Data unggahan gambar",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Judul Gambar",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Deskripsi Gambar",
                                        "type": "string"
                                    },
                                    "picture": {
                                        "description": "File Gambar",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/greet": {
            "get": {
                "tags": [
                    "greeting"
                ],
                "summary": "Returns a Sample API response",
                "description": "A sample greeting to test out the API",
                "operationId": "greet",
                "parameters": [
                    {
                        "name": "firstname",
                        "in": "query",
                        "description": "nama depan",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lastname",
                        "in": "query",
                        "description": "nama belakang",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        }
    }
}